bug fixes cache size and new hash function implemented

+++ src/gmpcachingxx.cpp
@@ -3,6 +3,6 @@
 namespace gmpcaching{
 MasterCache* globalcache;
 #ifdef GMPCACHING_CACHESIZE
-int globalcache_size = GMPCACHING_CACHESIZE;
+uint64_t globalcache_size = GMPCACHING_CACHESIZE;
 #endif
 }
